requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-SQLAlchemy
psycopg2
flask-wtf
WTForms
bootstrap
flask-uploads
datetime
# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gunicorn
